requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
blinker==1.7.0
python-dotenv==1.0.0

# Security - Password Hashing
bcrypt==4.1.2
Werkzeug==3.0.1

# Security - OTP/2FA
pyotp==2.9.0

# Security - QR Code Generation  
Pillow==10.1.0

# Security - Encryption
cryptography==41.0.7

# Email validation
email-validator==2.1.0

# Rate limiting
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dateutil==2.8.2
